feat(core): add print-access-token auth subcommand for ease of DX and piping to other CLI tools
