+++ docs/man/auth/print-access-token.md
@@ -1,0 +1,8 @@
+---
+title: Print the cached OIDC access token (if found)
+
+command:
+  name: print-access-token
+---
+
+Retrieves the cached OIDC token from the OS-specific keychain and prints if found.
